dev: remove scrollreadyset, always stdout
