+++ examples/aliases.rs
@@ -70,11 +70,10 @@ fn main() {
                 1.0 / 60.0,
             ))),
             bevy::input::InputPlugin::default(),
-            bevy_ratatui::RatatuiPlugins::default(),
             ReplPlugins,
         ))
         .add_repl_command::<SayCommand>()
         .add_observer(on_say)
-        .add_systems(PostStartup, instructions.after(ScrollRegionReadySet))
+        .add_systems(PostStartup, instructions)
         .run();
 }
