+++ examples/builtins.rs
@@ -63,7 +63,6 @@ fn main() {
             MinimalPlugins
                 .set(ScheduleRunnerPlugin::run_loop(Duration::from_secs_f64(1.0 / 60.0))),
             bevy::input::InputPlugin::default(),
-            bevy_ratatui::RatatuiPlugins::default(),
             ReplPlugins,
         ))
         .add_systems(PostStartup, instructions)
