+++ examples/custom_renderer.rs
@@ -75,7 +75,6 @@ fn main() {
                 60.0_f64.recip(),
             ))),
             bevy::input::InputPlugin::default(),
-            bevy_ratatui::RatatuiPlugins::default(),
             ReplPlugins.set(PromptPlugin {
                 renderer: Arc::new(CustomRenderer),
                 ..default()
@@ -83,6 +82,6 @@ fn main() {
         ))
         .add_repl_command::<PingCommand>()
         .add_observer(on_ping)
-        .add_systems(PostStartup, instructions.after(ScrollRegionReadySet))
+        .add_systems(PostStartup, instructions)
         .run();
 }
