+++ examples/default.rs
@@ -96,7 +96,6 @@ fn main() {
                 .set(ScheduleRunnerPlugin::run_loop(Duration::from_secs_f64(
                     1.0 / 60.0,
                 ))),
-            bevy_ratatui::RatatuiPlugins::default(),
             ReplPlugins,
         ))
         .add_repl_command::<PrintCommand>()
