+++ examples/fixed_update.rs
@@ -157,7 +157,6 @@ fn main() {
                 ))),
             bevy::input::InputPlugin::default(),
             bevy::log::LogPlugin::default(),
-            bevy_ratatui::RatatuiPlugins::default(),
             ReplPlugins,
         ))
         .init_resource::<TimerResource>()
@@ -169,6 +168,6 @@ fn main() {
         .add_observer(on_reset_timer)
         .add_systems(Update, display_timer_status.run_if(timer_is_running))
         .add_systems(FixedUpdate, update_timer)
-        .add_systems(PostStartup, instructions.after(ScrollRegionReadySet))
+        .add_systems(PostStartup, instructions)
         .run();
 }
