+++ examples/minimal.rs
@@ -49,6 +49,6 @@ fn main() {
         ))
         .add_repl_command::<PingCommand>()
         .add_observer(on_ping)
-        .add_systems(PostStartup, instructions.after(ScrollRegionReadySet))
+        .add_systems(PostStartup, instructions)
         .run();
 }
