+++ examples/update_resource.rs
@@ -125,7 +125,6 @@ fn main() {
             MinimalPlugins.set(ScheduleRunnerPlugin::run_loop(Duration::from_secs_f64(1.0 / 60.0))),
             // Required so the REPL can handle keyboard input
             bevy::input::InputPlugin::default(),
-            bevy_ratatui::RatatuiPlugins::default(),
             ReplPlugins,
         ))
         .add_repl_command::<TimeScaleCommand>()
